pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "medasr"
version = "0.1.0"
description = "Local voice dictation app using Google MedASR"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24.0",
    "sounddevice>=0.4.6",
    "PyQt6>=6.5.0",
    "pynput>=1.7.6",
    "pyyaml>=6.0",
    "torch>=2.0.0",
    "librosa>=0.10.0",
]

[project.scripts]
medasr = "medasr.__main__:main"
